+++ pkg/sentry/devices/tpuproxy/vfio/vfio.go
@@ -38,8 +38,6 @@ const (
 	tpuDeviceGroupName  = "vfio"
 	vfioDeviceGroupName = "vfio"
 
-	// VFIOPath is the valid path to a VFIO device.
-	VFIOPath = "/dev/vfio/vfio"
 )
 
 var (
