+++ runsc/cmd/gofer.go
@@ -16,23 +16,18 @@ package cmd
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
-	"regexp"
 	"runtime"
 	"runtime/debug"
-	"strings"
 
 	"github.com/google/subcommands"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"golang.org/x/sys/unix"
 	"gvisor.dev/gvisor/pkg/log"
-	"gvisor.dev/gvisor/pkg/sentry/devices/tpuproxy/vfio"
 	"gvisor.dev/gvisor/pkg/unet"
 	"gvisor.dev/gvisor/pkg/urpc"
-	"gvisor.dev/gvisor/runsc/boot"
 	"gvisor.dev/gvisor/runsc/cmd/sandboxsetup"
 	"gvisor.dev/gvisor/runsc/cmd/util"
 	"gvisor.dev/gvisor/runsc/config"
@@ -104,7 +99,7 @@ type Gofer struct {
 	devIoFD    int
 	applyCaps  bool
 	setUpRoot  bool
-	mountConfs boot.GoferMountConfFlags
+	mountConfs specutils.GoferMountConfFlags
 
 	// uid and gid are the user and group IDs to switch to after setting up the
 	// user namespace.
@@ -190,7 +185,7 @@ func (g *Gofer) Execute(_ context.Context, f *flag.FlagSet, args ...any) subcomm
 	defer goferToHostRPC.Close()
 
 	if g.setUpRoot {
-		if err := g.setupRootFS(spec, conf, goferToHostRPC); err != nil {
+		if err := sandboxsetup.SetupRootFS(spec, conf, g.mountConfs, g.devIoFD, makeRPCMountOpener(goferToHostRPC)); err != nil {
 			util.Fatalf("Error setting up root FS: %v", err)
 		}
 		if !conf.TestOnlyAllowRunAsCurrentUserWithoutChroot {
@@ -248,13 +243,13 @@ func (g *Gofer) Execute(_ context.Context, f *flag.FlagSet, args ...any) subcomm
 	//
 	// Note that all mount points have been mounted in the proper location in
 	// setupRootFS().
-	cleanMounts, err := g.resolveMounts(conf, spec.Mounts, root)
+	cleanMounts, err := sandboxsetup.ResolveMounts(conf, spec.Mounts, root, g.mountConfs)
 	if err != nil {
 		util.Fatalf("Failure to resolve mounts: %v", err)
 	}
 	spec.Mounts = cleanMounts
 	go func() {
-		if err := g.writeMounts(cleanMounts); err != nil {
+		if err := sandboxsetup.WriteMounts(g.mountsFD, cleanMounts); err != nil {
 			panic(fmt.Sprintf("Failed to write mounts: %v", err))
 		}
 	}()
@@ -265,7 +260,7 @@ func (g *Gofer) Execute(_ context.Context, f *flag.FlagSet, args ...any) subcomm
 	// modes exactly as sent by the sandbox, which will have applied its own umask.
 	unix.Umask(0)
 
-	procFDPath := procFDBindMount
+	procFDPath := sandboxsetup.ProcFDBindMount
 	if conf.TestOnlyAllowRunAsCurrentUserWithoutChroot {
 		procFDPath = "/proc/self/fd"
 	}
@@ -305,14 +300,6 @@ func (g *Gofer) Execute(_ context.Context, f *flag.FlagSet, args ...any) subcomm
 	return g.serve(spec, conf, root, ruid, euid, rgid, egid)
 }
 
-func newSocket(ioFD int) *unet.Socket {
-	socket, err := unet.NewSocket(ioFD)
-	if err != nil {
-		util.Fatalf("creating server on FD %d: %v", ioFD, err)
-	}
-	return socket
-}
-
 func (g *Gofer) serve(spec *specs.Spec, conf *config.Config, root string, ruid int, euid int, rgid int, egid int) subcommands.ExitStatus {
 	type connectionConfig struct {
 		sock      *unet.Socket
@@ -337,7 +324,7 @@ func (g *Gofer) serve(spec *specs.Spec, conf *config.Config, root string, ruid i
 	if rootfsConf.ShouldUseLisafs() {
 		// Start with root mount, then add any other additional mount as needed.
 		cfgs = append(cfgs, connectionConfig{
-			sock:      newSocket(ioFDs[0]),
+			sock:      sandboxsetup.NewSocket(ioFDs[0]),
 			mountPath: "/", // fsgofer process is always chroot()ed. So serve root.
 			readonly:  spec.Root.Readonly || rootfsConf.ShouldUseOverlayfs(),
 		})
@@ -366,7 +353,7 @@ func (g *Gofer) serve(spec *specs.Spec, conf *config.Config, root string, ruid i
 		ioFDs = ioFDs[1:]
 		readonly := specutils.IsReadonlyMount(m.Options) || mountConf.ShouldUseOverlayfs()
 		cfgs = append(cfgs, connectionConfig{
-			sock:      newSocket(ioFD),
+			sock:      sandboxsetup.NewSocket(ioFD),
 			mountPath: m.Destination,
 			readonly:  readonly,
 		})
@@ -379,7 +366,7 @@ func (g *Gofer) serve(spec *specs.Spec, conf *config.Config, root string, ruid i
 
 	if g.devIoFD >= 0 {
 		cfgs = append(cfgs, connectionConfig{
-			sock:      newSocket(g.devIoFD),
+			sock:      sandboxsetup.NewSocket(g.devIoFD),
 			mountPath: "/dev",
 		})
 		log.Infof("Serving /dev mapped on FD %d (ro: false)", g.devIoFD)
@@ -401,330 +388,16 @@ func (g *Gofer) serve(spec *specs.Spec, conf *config.Config, root string, ruid i
 	return subcommands.ExitSuccess
 }
 
-func (g *Gofer) writeMounts(mounts []specs.Mount) error {
-	bytes, err := json.Marshal(mounts)
-	if err != nil {
-		return err
+// makeRPCMountOpener returns a MountOpener that opens mount sources via the
+// gofer-to-host RPC channel.
+func makeRPCMountOpener(goferToHostRPC *urpc.Client) sandboxsetup.MountOpener {
+	return func(m *specs.Mount) (*os.File, error) {
+		var res container.OpenMountResult
+		if err := goferToHostRPC.Call("goferToHostRPC.OpenMount", m, &res); err != nil {
+			return nil, fmt.Errorf("opening %s: %w", m.Source, err)
+		}
+		return res.Files[0], nil
 	}
-
-	f := os.NewFile(uintptr(g.mountsFD), "mounts file")
-	defer f.Close()
-
-	for written := 0; written < len(bytes); {
-		w, err := f.Write(bytes[written:])
-		if err != nil {
-			return err
-		}
-		written += w
-	}
-	return nil
-}
-
-// Redhat distros don't allow to create bind-mounts in /proc/self directories.
-// It is protected by selinux rules.
-const procFDBindMount = "/proc/fs"
-
-func (g *Gofer) setupRootFS(spec *specs.Spec, conf *config.Config, goferToHostRPC *urpc.Client) error {
-	// Convert all shared mounts into slaves to be sure that nothing will be
-	// propagated outside of our namespace.
-	procPath := "/proc"
-	if err := specutils.SafeMount("", "/", "", unix.MS_SLAVE|unix.MS_REC, "", procPath); err != nil {
-		util.Fatalf("error converting mounts: %v", err)
-	}
-
-	root := spec.Root.Path
-	if !conf.TestOnlyAllowRunAsCurrentUserWithoutChroot {
-		// runsc can't be re-executed without /proc, so we create a tmpfs mount,
-		// mount ./proc and ./root there, then move this mount to the root and after
-		// setCapsAndCallSelf, runsc will chroot into /root.
-		//
-		// We need a directory to construct a new root and we know that
-		// runsc can't start without /proc, so we can use it for this.
-		flags := uintptr(unix.MS_NOSUID | unix.MS_NODEV | unix.MS_NOEXEC)
-		if err := specutils.SafeMount("runsc-root", "/proc/fs", "tmpfs", flags, "", procPath); err != nil {
-			util.Fatalf("error mounting tmpfs: %v", err)
-		}
-		if err := unix.Mount("", "/proc/fs", "", unix.MS_UNBINDABLE, ""); err != nil {
-			util.Fatalf("error setting MS_UNBINDABLE")
-		}
-		// Prepare tree structure for pivot_root(2).
-		if err := os.Mkdir("/proc/fs/proc", 0755); err != nil {
-			util.Fatalf("error creating /proc/fs/proc: %v", err)
-		}
-		if err := os.Mkdir("/proc/fs/root", 0755); err != nil {
-			util.Fatalf("error creating /proc/fs/root: %v", err)
-		}
-		if err := os.Mkdir("/proc/fs/etc", 0755); err != nil {
-			util.Fatalf("error creating /proc/fs/etc: %v", err)
-		}
-		// This cannot use SafeMount because there's no available procfs. But we
-		// know that /proc/fs is an empty tmpfs mount, so this is safe.
-		if err := unix.Mount("/proc", "/proc/fs/proc", "", flags|unix.MS_RDONLY|unix.MS_BIND|unix.MS_REC, ""); err != nil {
-			util.Fatalf("error mounting /proc/fs/proc: %v", err)
-		}
-		// self/fd is bind-mounted, so that the FD return by
-		// OpenProcSelfFD() does not allow escapes with walking ".." .
-		if err := unix.Mount("/proc/fs/proc/self/fd", "/proc/fs/"+procFDBindMount,
-			"", unix.MS_RDONLY|unix.MS_BIND|flags, ""); err != nil {
-			util.Fatalf("error mounting proc/self/fd: %v", err)
-		}
-		if err := sandboxsetup.CopyFile("/proc/fs/etc/localtime", "/etc/localtime"); err != nil {
-			log.Warningf("Failed to copy /etc/localtime: %v. UTC timezone will be used.", err)
-		}
-		root = "/proc/fs/root"
-		procPath = "/proc/fs/proc"
-	}
-
-	rootfsConf := g.mountConfs[0]
-	if rootfsConf.ShouldUseLisafs() {
-		// Mount root path followed by submounts.
-		if err := specutils.SafeMount(spec.Root.Path, root, "bind", unix.MS_BIND|unix.MS_REC, "", procPath); err != nil {
-			return fmt.Errorf("mounting root on root (%q) err: %v", root, err)
-		}
-
-		flags := uint32(unix.MS_SLAVE | unix.MS_REC)
-		if spec.Linux != nil && spec.Linux.RootfsPropagation != "" {
-			flags = specutils.PropOptionsToFlags([]string{spec.Linux.RootfsPropagation})
-		}
-		if err := specutils.SafeMount("", root, "", uintptr(flags), "", procPath); err != nil {
-			return fmt.Errorf("mounting root (%q) with flags: %#x, err: %v", root, flags, err)
-		}
-	}
-
-	// Replace the current spec, with the clean spec with symlinks resolved.
-	if err := g.setupMounts(conf, spec.Mounts, root, procPath, goferToHostRPC); err != nil {
-		util.Fatalf("error setting up FS: %v", err)
-	}
-
-	// Set up /dev directory is needed.
-	if g.devIoFD >= 0 {
-		g.setupDev(spec, conf, root, procPath)
-	}
-
-	// Check if root needs to be remounted as readonly.
-	if rootfsConf.ShouldUseLisafs() && (spec.Root.Readonly || rootfsConf.ShouldUseOverlayfs()) {
-		// If root is a mount point but not read-only, we can change mount options
-		// to make it read-only for extra safety.
-		// unix.MS_NOSUID and unix.MS_NODEV are included here not only
-		// for safety reasons but also because they can be locked and
-		// any attempts to unset them will fail.  See
-		// mount_namespaces(7) for more details.
-		log.Infof("Remounting root as readonly: %q", root)
-		flags := uintptr(unix.MS_BIND | unix.MS_REMOUNT | unix.MS_RDONLY | unix.MS_NOSUID | unix.MS_NODEV)
-		if err := specutils.SafeMount(root, root, "bind", flags, "", procPath); err != nil {
-			return fmt.Errorf("remounting root as read-only with source: %q, target: %q, flags: %#x, err: %v", root, root, flags, err)
-		}
-	}
-
-	if !conf.TestOnlyAllowRunAsCurrentUserWithoutChroot {
-		if err := sandboxsetup.PivotRoot("/proc/fs"); err != nil {
-			util.Fatalf("failed to change the root file system: %v", err)
-		}
-		if err := os.Chdir("/"); err != nil {
-			util.Fatalf("failed to change working directory")
-		}
-	}
-	return nil
-}
-
-// setupMounts bind mounts all mounts specified in the spec in their correct
-// location inside root. It will resolve relative paths and symlinks. It also
-// creates directories as needed.
-func (g *Gofer) setupMounts(conf *config.Config, mounts []specs.Mount, root, procPath string, goferToHostRPC *urpc.Client) (retErr error) {
-	mountIdx := 1 // First index is for rootfs.
-	for _, m := range mounts {
-		if !specutils.HasMountConfig(m) {
-			continue
-		}
-		mountConf := g.mountConfs[mountIdx]
-		mountIdx++
-		if !mountConf.ShouldUseLisafs() {
-			continue
-		}
-
-		dst, err := sandboxsetup.ResolveSymlinks(root, m.Destination)
-		if err != nil {
-			return fmt.Errorf("resolving symlinks to %q: %v", m.Destination, err)
-		}
-
-		flags := specutils.OptionsToFlags(m.Options) | unix.MS_BIND
-		if mountConf.ShouldUseOverlayfs() {
-			// Force mount read-only if writes are not going to be sent to it.
-			flags |= unix.MS_RDONLY
-		}
-
-		log.Infof("Mounting src: %q, dst: %q, flags: %#x", m.Source, dst, flags)
-		src := m.Source
-		var srcFile *os.File
-		if err := unix.Access(src, unix.R_OK); err != nil {
-			// The current process doesn't have enough permissions
-			// to open the mount, so let's try to open it in the
-			// parent user namespace.
-			var res container.OpenMountResult
-			if err := goferToHostRPC.Call("goferToHostRPC.OpenMount", &m, &res); err != nil {
-				return fmt.Errorf("opening %s: %w", m.Source, err)
-			}
-			srcFile = res.Files[0]
-			src = fmt.Sprintf("%s/self/fd/%d", procPath, srcFile.Fd())
-		}
-		err = specutils.SafeSetupAndMount(src, dst, m.Type, flags, procPath)
-		srcFile.Close()
-		if err != nil {
-			return fmt.Errorf("mounting %+v: %v", m, err)
-		}
-
-		dstFD, err := unix.Open(dst, unix.O_PATH|unix.O_CLOEXEC, 0)
-		if err != nil {
-			return fmt.Errorf("Open(%s, _, _): %w", dst, err)
-		}
-		defer unix.Close(dstFD)
-		// Apply mount options after creating all mount points.
-		// Otherwise they can be remounted into read-only.
-		defer func(dstFD int, flags uint32, dst string) {
-			path := fmt.Sprintf("/proc/self/fd/%d", dstFD)
-			// The gofer process doesn't execute anything natively.
-			flags |= unix.MS_NOSUID
-
-			statfs := unix.Statfs_t{}
-			if err := unix.Statfs(path, &statfs); err != nil {
-				retErr = fmt.Errorf("stat dst: %q", dst)
-				return
-			}
-			lockedFlags := uint32(0)
-			for _, f := range []struct {
-				st, ms int
-			}{
-				// MS_NOSUID are always set.
-				{unix.ST_RDONLY, unix.MS_RDONLY},
-				{unix.ST_NOEXEC, unix.MS_NOEXEC},
-				{unix.ST_NODEV, unix.MS_NODEV},
-				{unix.ST_NOATIME, unix.MS_NOATIME},
-				{unix.ST_NODIRATIME, unix.MS_NODIRATIME},
-				{unix.ST_RELATIME, unix.MS_RELATIME},
-			} {
-				if int(statfs.Flags)&f.st == f.st {
-					lockedFlags |= uint32(f.ms)
-				}
-			}
-			if lockedFlags&unix.MS_NOATIME|unix.MS_RELATIME == 0 {
-				lockedFlags |= unix.MS_STRICTATIME
-			}
-
-			// The previous SafeSetupAndMount creates a new bind-mount, but
-			// it doesn't change mount flags. A separate MS_BIND|MS_REMOUNT
-			// has to be done to apply the mount options.
-			if err := unix.Mount("", path, "", uintptr(flags|lockedFlags|unix.MS_REMOUNT), ""); err != nil {
-				retErr = fmt.Errorf("mount dst: %q, flags: %#x, err: %v", dst, flags, err)
-				return
-			}
-		}(dstFD, flags, dst)
-
-		// Set propagation options that cannot be set together with other options.
-		flags = specutils.PropOptionsToFlags(m.Options)
-		if flags != 0 {
-			if err := specutils.SafeMount("", dst, "", uintptr(flags), "", procPath); err != nil {
-				return fmt.Errorf("mount dst: %q, flags: %#x, err: %v", dst, flags, err)
-			}
-		}
-	}
-	return nil
-}
-
-// shouldExposeNvidiaDevice returns true if path refers to an Nvidia device
-// which should be exposed to the container.
-//
-// Precondition: nvproxy is enabled.
-func shouldExposeNvidiaDevice(path string) bool {
-	if !strings.HasPrefix(path, "/dev/nvidia") {
-		return false
-	}
-	if path == "/dev/nvidiactl" || path == "/dev/nvidia-uvm" {
-		return true
-	}
-	nvidiaDevPathReg := regexp.MustCompile(`^/dev/nvidia(\d+)$`)
-	return nvidiaDevPathReg.MatchString(path)
-}
-
-// shouldExposeVfioDevice returns true if path refers to an VFIO device
-// which should be exposed to the container.
-func shouldExposeVFIODevice(path string) bool {
-	return strings.HasPrefix(path, filepath.Dir(vfio.VFIOPath))
-}
-
-// shouldExposeTpuDevice returns true if path refers to a TPU device which
-// should be exposed to the container.
-//
-// Precondition: tpuproxy is enabled.
-func shouldExposeTpuDevice(path string) bool {
-	valid, _ := util.IsTPUDeviceValid(path)
-	return valid || shouldExposeVFIODevice(path)
-}
-
-func (g *Gofer) setupDev(spec *specs.Spec, conf *config.Config, root, procPath string) error {
-	if err := os.MkdirAll(filepath.Join(root, "dev"), 0777); err != nil {
-		return fmt.Errorf("creating dev directory: %v", err)
-	}
-	// Mount any devices specified in the spec.
-	if spec.Linux == nil {
-		return nil
-	}
-	nvproxyEnabled := specutils.NVProxyEnabled(spec, conf)
-	tpuproxyEnabled := specutils.TPUProxyIsEnabled(spec, conf)
-	for _, dev := range spec.Linux.Devices {
-		shouldMount := (nvproxyEnabled && shouldExposeNvidiaDevice(dev.Path)) ||
-			(tpuproxyEnabled && shouldExposeTpuDevice(dev.Path))
-		if !shouldMount {
-			continue
-		}
-		dst := filepath.Join(root, dev.Path)
-		log.Infof("Mounting device %q as bind mount at %q", dev.Path, dst)
-		if err := specutils.SafeSetupAndMount(dev.Path, dst, "bind", unix.MS_BIND, procPath); err != nil {
-			return fmt.Errorf("mounting %q: %v", dev.Path, err)
-		}
-	}
-	return nil
-}
-
-// resolveMounts resolved relative paths and symlinks to mount points.
-//
-// Note: mount points must already be in place for resolution to work.
-// Otherwise, it may follow symlinks to locations that would be overwritten
-// with another mount point and return the wrong location. In short, make sure
-// setupMounts() has been called before.
-func (g *Gofer) resolveMounts(conf *config.Config, mounts []specs.Mount, root string) ([]specs.Mount, error) {
-	mountIdx := 1 // First index is for rootfs.
-	cleanMounts := make([]specs.Mount, 0, len(mounts))
-	for _, m := range mounts {
-		if !specutils.HasMountConfig(m) {
-			cleanMounts = append(cleanMounts, m)
-			continue
-		}
-		mountConf := g.mountConfs[mountIdx]
-		mountIdx++
-		if !mountConf.ShouldUseLisafs() {
-			cleanMounts = append(cleanMounts, m)
-			continue
-		}
-		dst, err := sandboxsetup.ResolveSymlinks(root, m.Destination)
-		if err != nil {
-			return nil, fmt.Errorf("resolving symlinks to %q: %v", m.Destination, err)
-		}
-		relDst, err := filepath.Rel(root, dst)
-		if err != nil {
-			panic(fmt.Sprintf("%q could not be made relative to %q: %v", dst, root, err))
-		}
-
-		opts, err := sandboxsetup.AdjustMountOptions(conf, filepath.Join(root, relDst), m.Options)
-		if err != nil {
-			return nil, err
-		}
-
-		cpy := m
-		cpy.Destination = filepath.Join("/", relDst)
-		cpy.Options = opts
-		cleanMounts = append(cleanMounts, cpy)
-	}
-	return cleanMounts, nil
 }
 
 // setFlags sets sync FD flags on the given FlagSet.
