+++ runsc/specutils/BUILD
@@ -10,6 +10,7 @@ go_library(
     srcs = [
         "cri.go",
         "fs.go",
+        "gofer_conf.go",
         "namespace.go",
         "nvidia.go",
         "restore.go",
@@ -35,7 +36,10 @@ go_library(
 go_test(
     name = "specutils_test",
     size = "small",
-    srcs = ["specutils_test.go"],
+    srcs = [
+        "gofer_conf_test.go",
+        "specutils_test.go",
+    ],
     library = ":specutils",
     deps = [
         "//pkg/sentry/devices/nvproxy/nvconf",
